+++ source/chip/STM32/STM32F4/STM32F4-lowLevelInitialization.cpp
@@ -17,8 +17,6 @@
 #include "distortos/chip/STM32F4-RCC.hpp"
 #include "distortos/chip/STM32F4-RCC-bits.h"
 
-#include "distortos/architecture/ARMv6-M-ARMv7-M-configureSysTick.hpp"
-
 #include "distortos/chip/CMSIS-proxy.h"
 
 namespace distortos
@@ -142,14 +140,6 @@ void lowLevelInitialization()
 	switchSystemClock(systemClockSource);
 
 #endif	// def CONFIG_CHIP_STM32F4_STANDARD_CLOCK_CONFIGURATION_ENABLE
-
-	constexpr uint32_t period {ahbFrequency / CONFIG_TICK_FREQUENCY};
-	constexpr uint32_t periodDividedBy8 {period / 8};
-	constexpr bool divideBy8 {period > architecture::maxSysTickPeriod};
-	// at least one of the periods must be valid
-	static_assert(period <= architecture::maxSysTickPeriod || periodDividedBy8 <= architecture::maxSysTickPeriod,
-			"Invalid SysTick configuration!");
-	architecture::configureSysTick(divideBy8 == false ? period : periodDividedBy8, divideBy8);
 }
 
 }	// namespace chip